requirements.txt
amqp==5.1.1
asgiref==3.5.1
async-timeout==4.0.2
billiard==3.6.4.0
black==22.3.0
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Deprecated==1.2.13
Django==3.2
django-crontab==0.7.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
dnspython==2.2.1
eventlet==0.33.1
feedparser==6.0.8
flake8==4.0.1
greenlet==2.0.0a2
idna==3.3
importlib-metadata==4.11.3
kombu==5.2.4
Markdown==3.3.7
mccabe==0.6.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
prompt-toolkit==3.0.29
punq==0.6.2
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.1
requests==2.27.1
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.0
